pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "igcraft"
version = "0.0.1"
description = "IgCraft: paired human antibody sequence design."
authors = [
  { name = "Matt Greenig", email = "mg989@cam.ac.uk" },
  { name = "Haowen Zhao", email = "hz362@cam.ac.uk" }
]
requires-python = ">=3.12"
dependencies = [
  "biopython==1.84",
  "pandas==2.2.2",
  "pydantic==2.9.2",
  "numpy==2.1.2",
  "torch>=2.4",
  "pytorch-lightning==2.4.0",
  "einops==0.8.0",
  "hydra-core==1.3.2",
  "wandb==0.18.5",
  "loguru==0.7.2",
  "h5py==3.12.1",
  "fast-edit-distance==1.2.2",
  "polars-lts-cpu==1.12.0"
]

[tool.setuptools.packages.find]
where = ["."]